+++ src/entity/Student.ts
@@ -4,7 +4,7 @@ import { Family } from "./Family";
 
 @Entity("students")
 export class Student {
-    @PrimaryGeneratedColumn({ type: "int", unsigned: true })
+    @PrimaryGeneratedColumn({ type: "bigint", unsigned: true })
     id: number;
 
     @Column({ type: "varchar", length: 255 })
